detect and delete double IDs
We have 2 traffics which have the same IDs from time to
time as FlyingBird22 reported in
https://gitter.im/FlyingBird22. It is caused by
"the newly received aircraft information is stored in the
slot which is older than ENTRY_EXPIRATION_TIME".

Let's assume the aircraft A is recorded in slot 7.
The aircraft A transmits its new location. The new location
may be recorded in slot 2 which is expired and used to have
aircraft B. Now the same aircraft exist in slot 2 and 7.

Signed-off-by: caz yokoyama <caz@caztech.com>